Fix bug on Linux where hot reload in src/main file froze window
When a file in `src/main` is changed the `vite-plugin-electron` will
kill the electron process and restart it. However, on Linux it seems
like the electron window process freezes if we do that.

Killing the main electron process seems to kill all its children on
Linux, but killing the children before the electron process seems to
freeze the window. In order to fix this bug we only kill the electron
process.

We will continue to manually kill the process children before the
parent process on macOS, as the current implementation seems to work
well on that platform.

+++ desktop/packages/mullvad-vpn/vite-utils.ts
@@ -41,9 +41,11 @@ function pidTree(tree: PidTree) {
 }
 
 function killTree(tree: PidTree) {
-  if (tree.children) {
-    for (const child of tree.children) {
-      killTree(child);
+  if (process.platform === 'darwin') {
+    if (tree.children) {
+      for (const child of tree.children) {
+        killTree(child);
+      }
     }
   }
